Resolution Setting Added - Needs Work
Functionality of changing resolution and window mode is working minus a few errors.

For example the combo box to select a window mode only ever returns -1 which means whatever option is pressed full-screen mode is applied.

Also when windowed the resolutions appear to be anywhere between 4x-8x larger in terms of space on screen than they should be.